refactor(ramp): move Ramp index to root folder
